Added Header, Home, Login, SignUp,Upload Components
Header consists of the sliding animation which on clicked gives options to navigate through, Home is the where all the slideshow of images and Services are mentioned. Login page enables users to login and sign up does for a new user in this website. Upload enables the user to upload videos in the website.